fix: support for python 3.x
current dode doesn't support pyton 3 and breaks by doing so.
This is because the command used is python and python on most linux computers are
python 2.7 by default.

Overriding this on users computer would break a lot of inbult programs that
depends on version 2.7 on linux.

So, the best is to have the command changed to python3 instead of python
this makes sense as alot of people now code in python3.x and it makes it
easier when deploying, instead of always changing server config to
override the default python3 that is inbuilt

Resolves: #80